refactor(react): cleanup provider
